feat: add http3 support, static file serving seems to be broken?

+++ shell.nix
@@ -8,41 +8,53 @@ with pkgs;
       openssl
       pkg-config
       curlHTTP3
+      certbot
     ];
 
     shellHook = ''
       mkdir -p .certs
 
-      if [[ ! -f .certs/localhost.crt ]] || [[ ! -f .certs/localhost.key ]]; then
-        openssl req -x509 -newkey rsa:2048 -nodes -keyout ./.certs/localhost.key \
-            -out ./.certs/localhost.crt -days 365 -subj "/CN=localhost" \
-            -addext "subjectAltName = DNS:localhost"
+      USE_SELF_SIGNED_CERT=false
+      USE_GENERATED_CERT=true
+
+      export USE_SELF_SIGNED_CERT
+      export USE_GENERATED_CERT
+
+      if [[ $USE_SELF_SIGNED_CERT == true ]]; then
+        if [[ ! -f .certs/localhost.crt ]] || [[ ! -f .certs/localhost.key ]]; then
+          openssl req -x509 -newkey rsa:2048 -nodes -keyout ./.certs/localhost.key \
+              -out ./.certs/localhost.crt -days 365 -subj "/CN=localhost" \
+              -addext "subjectAltName = DNS:localhost"
+        fi
+
+        local EXPIRATION_DATE=$(openssl x509 -in ./.certs/localhost.crt -noout -enddate | cut -d= -f2)
+        echo "Certificate expires on $EXPIRATION_DATE"
+        local EXPIRATION_EPOCH=$(date -d "$EXPIRATION_DATE" +%s)
+        local NOW_EPOCH=$(date +%s)
+
+        if [[ $EXPIRATION_EPOCH -lt $NOW_EPOCH ]]; then
+          echo "Certificate expired on $EXPIRATION_DATE"
+          echo "Certificate expired on $EXPIRATION_DATE"
+
+          # Generate new CSR
+          openssl x509 -x509toreq -in .certs/localhost.crt -out .certs/csr.pem -signkey .certs/localhost.key
+
+          # Generate new certificate with extended validity
+          openssl x509 -req -days 365 -in .certs/csr.pem -signkey .certs/localhost.key -out .certs/localhost.crt
+
+          # Clean up CSR
+          rm .certs/csr.pem
+        fi
+
+        cat .certs/localhost.crt /etc/ssl/certs/ca-certificates.crt > .certs/ca-bundle.crt
+
+        export SSL_CERT_FILE=.certs/ca-bundle.crt
+        export SSL_CERT_DIR=.certs
+        echo "Temporary certificate trust enabled"
       fi
 
-      local EXPIRATION_DATE=$(openssl x509 -in ./.certs/localhost.crt -noout -enddate | cut -d= -f2)
-      echo "Certificate expires on $EXPIRATION_DATE"
-      local EXPIRATION_EPOCH=$(date -d "$EXPIRATION_DATE" +%s)
-      local NOW_EPOCH=$(date +%s)
-
-      if [[ $EXPIRATION_EPOCH -lt $NOW_EPOCH ]]; then
-        echo "Certificate expired on $EXPIRATION_DATE"
-        echo "Renewing certificate..."
-
-        # Generate new CSR
-        openssl x509 -x509toreq -in .certs/localhost.crt -out .certs/csr.pem -signkey .certs/localhost.key
-
-        # Generate new certificate with extended validity
-        openssl x509 -req -days 365 -in .certs/csr.pem -signkey .certs/localhost.key -out .certs/localhost.crt
-
-        # Clean up CSR
-        rm .certs/csr.pem
+      if [[ $USE_GENERATED_CERT == true ]]; then
+        echo "Using generated certificate"
       fi
-
-      cat .certs/localhost.crt /etc/ssl/certs/ca-certificates.crt > .certs/ca-bundle.crt
-
-      export SSL_CERT_FILE=.certs/ca-bundle.crt
-      export SSL_CERT_DIR=.certs
-
-      echo "Temporary certificate trust enabled"
     '';
   }
